adjust time tool

+++ aaas-api-user-info/src/main/java/com/shellshellfish/aaas/userinfo/dao/repositories/mysql/UserInfoFriendRuleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 public interface UserInfoFriendRuleRepository extends PagingAndSortingRepository<UiFriendRule,
     Long> {
-  List<UiFriendRule> findAllByBankId( BigInteger bankId);
+  List<UiFriendRule> findAllByBankId( Long bankId);
   List<UiFriendRule> findAll();
 }
